+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Unreleased and unplanned todos
 Released: TBA.
 [Diff](https://github.com/kvz/scrolex/compare/v0.0.16...master).
 
+- [x] Add method: `failure`
 - [x] Document `env`
 
 ## v0.0.16